feat: add csrf protection

+++ src/Controller/AccountController.php
@@ -82,7 +82,7 @@ class AccountController extends AbstractController
         ValidatorInterface $validator,
         EntityManagerInterface $entityManager
     ): Response {
-        if ($request->isMethod('POST')) {
+        if ($request->isMethod('POST') && $this->isCsrfTokenValid('update-email', $request->request->get('token'))) {
             $subscriberEmailUpdate = new SubscriberEmailUpdate();
             $subscriberEmailUpdate->hydrateFromData($request->request->all());
 
@@ -112,7 +112,7 @@ class AccountController extends AbstractController
         UserPasswordHasherInterface $passwordHasher,
         EntityManagerInterface $entityManager
     ): Response {
-        if ($request->isMethod('POST')) {
+        if ($request->isMethod('POST') && $this->isCsrfTokenValid('update-password', $request->request->get('token'))) {
             $subscriber = $this->getUser();
 
             $subscriberPasswordUpdate = new SubscriberPasswordUpdate();
@@ -152,7 +152,7 @@ class AccountController extends AbstractController
         CityService $cityService,
         EntityManagerInterface $entityManager
     ): Response {
-        if ($request->isMethod('POST')) {
+        if ($request->isMethod('POST') && $this->isCsrfTokenValid('update-user-infos', $request->request->get('token'))) {
             $subscriberContactInfosUpdate = new SubscriberContactInfosUpdate();
             $subscriberContactInfosUpdate->hydrateFromData($request->request->all());
             $cityService->processCityDetails($subscriberContactInfosUpdate);
@@ -187,7 +187,7 @@ class AccountController extends AbstractController
         ValidatorInterface $validator,
         EntityManagerInterface $entityManager
     ): Response {
-        if ($request->isMethod('POST')) {
+        if ($request->isMethod('POST') && $this->isCsrfTokenValid('update-platforms', $request->request->get('token'))) {
             $subscriberStreamingPlatformsUpdate = new SubscriberStreamingPlatformsUpdate();
             $subscriberStreamingPlatformsUpdate->hydrateFromData($request->request->all());
 
@@ -222,7 +222,7 @@ class AccountController extends AbstractController
         ValidatorInterface $validator,
         EntityManagerInterface $entityManager
     ): Response {
-        if ($request->isMethod('POST')) {
+        if ($request->isMethod('POST') && $this->isCsrfTokenValid('add-child', $request->request->get('token'))) {
             $subscriberChildCreate = new SubscriberChildCreate();
             $subscriberChildCreate->hydrateFromData($request->request->all());
 
@@ -248,66 +248,72 @@ class AccountController extends AbstractController
         return $this->render('account/add_child.html.twig');
     }
 
-    #[Route('/supprimer/enfant/{id}', name: 'app_remove_child')]
+    #[Route('/supprimer/enfant/', name: 'app_remove_child')]
     public function removeChild(
         Request $request,
         ChildRepository $childRepository,
         EntityManagerInterface $entityManager,
-        int $id
     ): Response {
-        /**
-         * @var Subscriber $subscriber
-         */
-        $subscriber = $this->getUser();
 
-        $child = $childRepository->findOneBy([
-            'id' => $id,
-        ]);
+        if ($request->isMethod('POST') && $this->isCsrfTokenValid('remove-child', $request->request->get('token'))) {
+            /**
+             * @var Subscriber $subscriber
+             */
+            $subscriber = $this->getUser();
 
-        if ($child === null || $subscriber->getId() !== $child->getSubscriber()?->getId() || $subscriber->getChilds()->count() < 2) {
-            return $this->redirectToRoute('app_account');
+            $child = $childRepository->findOneBy([
+                'id' => $request->request->get('child_id'),
+            ]);
+
+
+            if ($child === null || $subscriber->getId() !== $child->getSubscriber()?->getId() || $subscriber->getChilds()->count() < 2) {
+                return $this->redirectToRoute('app_account');
+            }
+
+            $subscriber->removeChild($child);
+            $entityManager->remove($child);
+            $entityManager->flush();
+            $this->addFlash('success', "✅ C'est noté ! Tu recevras plus de recommandations de films pour {$child->getFirstname()}");
         }
-
-        $subscriber->removeChild($child);
-        $entityManager->remove($child);
-        $entityManager->flush();
-        $this->addFlash('success', "✅ C'est noté ! Tu recevras plus de recommandations de films pour {$child->getFirstname()}");
 
         return $this->redirectToRoute('app_account');
     }
 
-    #[Route('/supprimer/compte', name: 'app_remove_account', methods: ['POST'])]
+    #[Route('/supprimer/compte', name: 'app_remove_account')]
     public function removeAccount(
         Request $request,
         UserPasswordHasherInterface $passwordHasher,
         EntityManagerInterface $entityManager
     ): Response {
-        /**
-         * @var Subscriber $subscriber
-         */
-        $subscriber = $this->getUser();
-        $password   = $request->request->get('password');
 
-        /**
-         * @var PasswordAuthenticatedUserInterface $subscriber
-         */
-        $isPasswordValid = $passwordHasher->isPasswordValid(
-            $subscriber,
-            (string) $password
-        );
+        if ($request->isMethod('POST') && $this->isCsrfTokenValid('remove-account', $request->request->get('token'))) {
+            /**
+             * @var Subscriber $subscriber
+             */
+            $subscriber = $this->getUser();
+            $password   = $request->request->get('password');
 
-        /**
-         * @var Subscriber $subscriber
-         */
-        if ($isPasswordValid === true) {
-            $entityManager->remove($subscriber);
-            $entityManager->flush();
-            $session = new Session();
-            $session->invalidate();
+            /**
+             * @var PasswordAuthenticatedUserInterface $subscriber
+             */
+            $isPasswordValid = $passwordHasher->isPasswordValid(
+                $subscriber,
+                (string) $password
+            );
 
-            return $this->redirectToRoute('app_logout');
+            /**
+             * @var Subscriber $subscriber
+             */
+            if ($isPasswordValid === true) {
+                $entityManager->remove($subscriber);
+                $entityManager->flush();
+                $session = new Session();
+                $session->invalidate();
+
+                return $this->redirectToRoute('app_logout');
+            }
+            $this->addFlash('cant_remove_account', '');
         }
-        $this->addFlash('cant_remove_account', '');
 
         return $this->redirectToRoute('app_account');
     }
